+++ packages/keyring-snap-bridge/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- `SnapKeyringV1.setSelectedAccounts` only forwards account IDs that exist on this snap; unknown IDs are omitted and logged ([#TODO](https://github.com/MetaMask/accounts/pull/TODO))
+
 ## [21.0.1]
 
 ### Changed